+++ app/pages/activity/index.vue
@@ -1,0 +1,14 @@
+<script lang="ts" setup>
+definePageMeta({
+  middleware: ['auth'],
+  breadcrumb: '活动记录',
+})
+</script>
+
+<template>
+  <div />
+</template>
+
+<style>
+
+</style>